+++ src/fr/math/minecraft/shared/inventory/items/shovel/IronShovelCraft.java
@@ -13,7 +13,7 @@ public class IronShovelCraft extends CraftRecipes {
     }
     @Override
     public void fillRecipe() {
-        CraftData data1 = new CraftData(new int[]
+        CraftData data1 = new CraftData(new byte[]
                 {
                         Material.IRON_INGOT.getId(), Material.AIR.getId(), Material.AIR.getId(),
                         Material.STICK.getId(), Material.AIR.getId(), Material.AIR.getId(),
@@ -21,7 +21,7 @@ public class IronShovelCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data2 = new CraftData(new int[]
+        CraftData data2 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.IRON_INGOT.getId(), Material.AIR.getId(),
                         Material.AIR.getId(), Material.STICK.getId(), Material.AIR.getId(),
@@ -29,7 +29,7 @@ public class IronShovelCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data3 = new CraftData(new int[]
+        CraftData data3 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.AIR.getId(), Material.IRON_INGOT.getId(),
                         Material.AIR.getId(), Material.AIR.getId(), Material.STICK.getId(),
